+++ src/index.ts
@@ -9,16 +9,16 @@ export * from './resizable';
 export function configure(config, options) {
 
   let attributes = [
-    "./interact-draggable",
-    "./interact-dropzone",
-    "./interact-gesturable",
-    "./interact-resizable",
-    "./resizable"
+    './interact-draggable',
+    './interact-dropzone',
+    './interact-gesturable',
+    './interact-resizable',
+    './resizable',
   ];
 
   if (options && options.enableDragDropAttributes === true) {
-    attributes.push("./draggable");
-    attributes.push("./dropzone");
+    attributes.push('./draggable');
+    attributes.push('./dropzone');
   }
 
   config.globalResources(attributes);