+++ src/interact-draggable.ts
@@ -1,5 +1,5 @@
-import { inject } from "aurelia-framework";
-import * as interact from "interact";
+import { inject } from 'aurelia-framework';
+import * as interact from 'interact';
 
 @inject(Element)
 export class InteractDraggableCustomAttribute {
@@ -7,17 +7,17 @@ export class InteractDraggableCustomAttribute {
   /**
    * interact.js options
    */
-  value: {[key: string]: any};
+  public value: {[key: string]: any};
 
   constructor(private element: HTMLElement) { }
 
   public attached() {
     interact(this.element)
       .draggable(Object.assign({}, this.value || {}))
-      .on("dragstart", (event) => this.dispatch("interact-dragstart", event))
-      .on("dragmove", (event) => this.dispatch("interact-dragmove", event))
-      .on("draginertiastart", (event) => this.dispatch("interact-draginertiastart", event))
-      .on("dragend", (event) => this.dispatch("interact-dragend", event));
+      .on('dragstart', (event) => this.dispatch('interact-dragstart', event))
+      .on('dragmove', (event) => this.dispatch('interact-dragmove', event))
+      .on('draginertiastart', (event) => this.dispatch('interact-draginertiastart', event))
+      .on('dragend', (event) => this.dispatch('interact-dragend', event));
   }
 
   private dispatch(name, data) {