+++ src/interact-dropzone.ts
@@ -1,5 +1,5 @@
-import { inject } from "aurelia-framework";
-import * as interact from "interact";
+import { inject } from 'aurelia-framework';
+import * as interact from 'interact';
 
 @inject(Element)
 export class InteractDropzoneCustomAttribute {
@@ -7,18 +7,18 @@ export class InteractDropzoneCustomAttribute {
   /**
    * interact.js options
    */
-  value: {[key: string]: any};
+  public value: {[key: string]: any};
 
   constructor(private element: HTMLElement) { }
 
   public attached() {
     interact(this.element)
       .dropzone(Object.assign({}, this.value || {}))
-        .on("dropactivate", (event) => this.dispatch("interact-dropactivate", event))
-        .on("dragenter", (event) => this.dispatch("interact-dragenter", event))
-        .on("dragleave", (event) => this.dispatch("interact-dragleave", event))
-        .on("drop", (event) => this.dispatch("interact-drop", event))
-        .on("dropdeactivate", (event) => this.dispatch("interact-dropdeactivate", event));
+        .on('dropactivate', (event) => this.dispatch('interact-dropactivate', event))
+        .on('dragenter', (event) => this.dispatch('interact-dragenter', event))
+        .on('dragleave', (event) => this.dispatch('interact-dragleave', event))
+        .on('drop', (event) => this.dispatch('interact-drop', event))
+        .on('dropdeactivate', (event) => this.dispatch('interact-dropdeactivate', event));
   }
 
   private dispatch(name, data) {