+++ src/interact-gesturable.ts
@@ -1,5 +1,5 @@
-import { inject } from "aurelia-framework";
-import * as interact from "interact";
+import { inject } from 'aurelia-framework';
+import * as interact from 'interact';
 
 @inject(Element)
 export class InteractGesturableCustomAttribute {
@@ -7,16 +7,16 @@ export class InteractGesturableCustomAttribute {
   /**
    * interact.js options
    */
-  value: {[key: string]: any};
+  public value: {[key: string]: any};
 
   constructor(private element: HTMLElement) { }
 
   public attached() {
     interact(this.element)
       .dropzone(Object.assign({}, this.value || {}))
-        .on("gesturestart", (event) => this.dispatch("interact-gesturestart", event))
-        .on("gesturemove", (event) => this.dispatch("interact-gesturemove", event))
-        .on("gestureend", (event) => this.dispatch("interact-gestureend", event));
+        .on('gesturestart', (event) => this.dispatch('interact-gesturestart', event))
+        .on('gesturemove', (event) => this.dispatch('interact-gesturemove', event))
+        .on('gestureend', (event) => this.dispatch('interact-gestureend', event));
   }
 
   private dispatch(name, data) {