+++ doc/CHANGELOG.md
@@ -1,3 +1,27 @@
+<a name="1.0.11"></a>
+## [1.0.11](https://github.com/eriklieben/aurelia-interactjs/compare/v1.0.6...v1.0.11) (2016-10-20)
+
+
+### Bug Fixes
+
+* **interact:** import 'interact' instead of 'interact.js' and bump version ([0ad0795](https://github.com/eriklieben/aurelia-interactjs/commit/0ad0795))
+* **jspm:** add interact.js with interact alias ([e4045cb](https://github.com/eriklieben/aurelia-interactjs/commit/e4045cb))
+* **jspm:** disable jspm node convention ([72b66d7](https://github.com/eriklieben/aurelia-interactjs/commit/72b66d7))
+
+
+
+<a name="1.0.10"></a>
+## [1.0.10](https://github.com/eriklieben/aurelia-interactjs/compare/v1.0.6...v1.0.10) (2016-10-20)
+
+
+### Bug Fixes
+
+* **interact:** import 'interact' instead of 'interact.js' and bump version ([0ad0795](https://github.com/eriklieben/aurelia-interactjs/commit/0ad0795))
+* **jspm:** add interact.js with interact alias ([e4045cb](https://github.com/eriklieben/aurelia-interactjs/commit/e4045cb))
+* **jspm:** disable jspm node convention ([72b66d7](https://github.com/eriklieben/aurelia-interactjs/commit/72b66d7))
+
+
+
 <a name="1.0.9"></a>
 ## [1.0.9](https://github.com/eriklieben/aurelia-interactjs/compare/v1.0.6...v1.0.9) (2016-10-17)